// src/app/models/user.ts
export class User {
    Address: string;
    Birthday: string;
    Email: string;
    FullName: string;
    First: string;
    Last: string;
    Id: number;
    ImageUrl: string;
    LocationId: number;
    Phone: string;
    PositionId: number;
    Position: {};
    Projects: {}[];
    Roles: {}[];
    Skype: string;
  }
  